+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+### Added
+- support for using tuning parameter names in the problem size
 
 ## [0.0.1] - 2016-06-14
 ### Added